backend/requirements.txt
fastapi
uvicorn
pandas
numpy
requests
tldextract
scikit-learn
tensorflow
python-multipart
fastapi
uvicorn
pandas
requests
beautifulsoup4
numpy
scikit-learn
xgboost
tensorflow
python-whois